add test for hooks
